handler.rsから各moduleに処理を移動
