+++ src/main.rs
@@ -1,4 +1,6 @@
+mod file;
 mod handler;
+mod parser;
 
 fn main() {
     // TODO: catch panic
